Plutus.PAB.Run.PSGenerator

Documentation

class HasPSTypes a where

List of types linked to contract type a that need to be available in Purescript.

Minimal complete definition

Nothing

Methods

psTypes :: [SumType 'Haskell]

pabBridge :: BridgePart

PAB's main bridge that includes common bridges

data PabBridge

pabBridgeProxy :: Proxy PabBridge

pabTypes :: [SumType 'Haskell]

PAB's list of types that includes common types.

pabSettings :: Settings

generateAPIModule

Arguments

:: forall a. (Typeable a, HasPSTypes a)

=> Proxy a

-> FilePath
Output directory of PS files

-> IO ()

Use the Proxy for specifying a when generating PS functions for the webserver using a specific Builtin (ex. generate 'Builtin Marlowe' instead of 'Builtin a').